chore: move functions and classes out of main.py
