--- MANIFEST.in
@@ -1,10 +1,0 @@
-include AUTHORS.rst
-include LICENSE
-include MANIFEST.in
-include README.rst
-include test_docformatter.py
-recursive-include images *.png
-
-exclude .pre-commit-hooks.yaml
-exclude .travis.yml
-exclude Makefile
